requirements.txt
attrs==23.2.0
blinker==1.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
urllib3==2.2.1
Werkzeug==3.0.2
